Fixed vector index out of bound bug & Overloaded Mesh::init function in component
Vector index out of bound issue is caused by rand() function not using the correct value. Once that is fixed, I discovered that the rand() is using the same seed, so I had to initialize a random seed based on time. In addition, I also overloaded the Mesh::init function in component so that it can be used with spritesheet.