Tiles will be now only be rendered once
All the tiles will now be render to a texture on a custom frame buffer object. Once the tile is render onto the texture, the texture is saved in the world and it will no longer be modified. Later on, the game will render the whole texture every frame and everything else will be rendered on top of this texture. This allows us to only render to tile once.